Fixed The GitProject and Added a bunch of stuff
Commit

Scripts:
added a FPCamera mode to the playercontroller, compleate with mouse
movement and options for sensitivity.
Rewrote the firing system to let the playercontroller controll the
firing of the arrow.
Made the arrow behave more like a normal arrow in the air.
rewrote the way the way the game detects where the player wants to move,
fixing a problem with the navmesh.

Scene:
added a new scene: Prototype_CoreMechanics.

Prefabs:
Added the main camera as a prefab
Added the player as a prefab
Added A firstperson mode trigger

Misc
Had to copy over a bunch of files from the project folder i worked in
since it was the wrong one