FixOfTheCaracterController
Scripts:
-Major cleanup of the playercontroller, this will make it easier to
setup the player in unity. More needed though
-removes alot of unesecary public vars
-made alot of vars protected

-ArrowController (did some cleanup here too)
-made a variation of the FireArrow() function to now accept no
parameters an only fire with constant speed.
-broke the arrow compleatly dont know what is wrong but its the roation
bit thats mis behaving

-remade the cameracontroll yet again. it now controlls two separate
object for the X and Z axis.
-added a reset so that the camera does not maintain its rotation during
transitions from FPMode to TPMode and the other way around
-tried abit back and fourt to fix a problm where the arrow does not keep
its rotation when fired

Prefabs:
reamde the playerprefab yet again
deleated the old playerPrefab

Scenes:
created a new scene to setup the playerPrefab
-changes the scale on the buildings in the Prototype_allTogheterhers
scene

misc
fixed pretty much every problem i had with the playercontroller plus i
got to clean up alot of the code in the process
there is an error with the arrow, it does not keep its rotation when the
player fires it. It is always shot out paralell to the ground. It do
accept the rotation around the y axis but refuses the x axis.